#2405 SOAP error - added user soap-services and ROLE_SERVICES; corrected spring-security.xml, page.tag; User.attributes change HashMap on ConcurrentHashMap; AuthenticationUtils.authenticateLocal added remove roles; added V2_7_4_0__AddMonitoringAndSoapServicesUser and removed V2_7_4_0__MonitoringUser;
